Set specific exit code to indicate a parse error
This causes the exit code from wattsi to be set to 65 in the case where wattsi
has reported any parse errors. This is useful because parse errors are the only
case for which wattsi currently emits line/column numbers, and it's useful to
know if wattsi has reported any messages with line/column numbers.