Make the MODE="RELEASE" build less noisy
When running the `MODE="RELEASE"` build, this change causes fpc to suppress
a number of different notes, hints, and warnings that we probably don't
need to be exposing to normal contributors running the build on their own.
The messages remain available when running the `MODE="DEBUG"`, etc., builds
(in which case they get processed through the `src/lib/filter.pl` script).