Generate multipage <dfn> cross-reference data
This change causes wattsi to emit a new xrefs.json file as part of the
build. For each <dfn> element in the spec source, the xrefs.json file
has a list of the spec section names that have a cross-reference to that
<dfn>, and for each section name, a list of cross-reference anchors from
that section to the <dfn>.

This is part of https://github.com/whatwg/html/pull/2638.