Add IDs for unreferenced subdfns
In the developer's edition, previously if a subdfn was not ever
referenced, it would not get an ID. This meant it could not be linked to
from other pages on the web, which is bad, and also interfered with
trying to make the caniuse annotation code work, for whatwg/html#2794.
This change ensures that all subdfns get IDs.