Avoid linking to index.html
Make the Table of Contents link to ./ instead of index.html.

Make fragment-links.json store the empty string instead of index.html.
Having index.html was bogus to begin with, because link-fixup.js would
then redirect to index.html.html.

link-fixup.js needs to be updated for this change to work correctly.