Do not always ensure IDs
This is a partial revert of de9bac81b79ea722ede505ebf039c3e4ecb6063e.
This fixes https://github.com/whatwg/html-build/issues/121. Hopefully we
can work out something that both helps with the dev edition IDs and also
preserves IDs in the other editions, but for now, it's best to just
revert that.

This means that the dev edition will not always have IDs for the
"subdfns" included there, unless they are cross-referenced from
elsewhere. In turn, this means that the linkages to caniuse won't be
detectable in such cases.