Add procedure to save HTML fragment to string
This change adds a procedure useful in combination with the syntax-
highlighting output feature — specifically, in the case where the
highlighter fails to produce any output.

This change enables the serialized HTML of a node that will be passed to
the highlighter to be first saved as a string. So in the case where the
highlighter fails, the saved HTML string (unhighlighted) can be written
to the output instead of the (expected-but-missing) highlighter output.