Ensure IDL blocks are omitted from the dev edition
The existing wattsi code for causing IDL blocks to be omitted from the
dev edition quit working when https://github.com/whatwg/html/pull/3768
moved class=idl attributes off `pre` elements and instead onto to
`code` children of those elements,

So the first part of this change causes those `code class=idl` elements
to be dropped as expected. But as a result of that, the now-empty `pre`
element parents of those `code` elements are left behind. So the second
part of this change looks for those empty `pre` elements and drops them.

This change reverts the e1e3599 revert that was made because — as
initially implemented in c13e47b — stray `</pre>` end tags got left
behind for all `pre` elements that were omitted. This change corrects
that problem (it ensures the end tags get removed as expected).

Fixes https://github.com/whatwg/wattsi/issues/84