Allow <pre> to have <code> children w/o data-x=""
This change allows `pre` elements to have `code` children that don’t
reference a `dfn` and that don’t need to to have an explicit `data-x=""`
to indicate they’re not referencing a `dfn`.

Otherwise without this change, every `code` element in the source that’s
a child of a `pre` needs to have `data-x=""` — because lacking that,
wattsi will emit an error for it (with the message “missing `<dfn>` for
topic "foo" explicitly from `<code>` element containing "bar".”).

Fixes https://github.com/whatwg/wattsi/issues/78