Ensure IDL blocks are omitted from the dev edition
The existing wattsi code for causing IDL blocks to be omitted from the
dev edition quit working when https://github.com/whatwg/html/pull/3768
moved class=idl attributes off `pre` elements and instead onto to
`code` children of those elements.

So the first part of this change causes those `code class=idl` elements
to be dropped as expected. But as a result of that, the now-empty `pre`
element parents of those `code` elements are left behind. So the second
part of this change looks for those empty `pre` elements and drops them.

Fixes #84.