Insert MDN annos for <p> descendants after the <p>
This change causes the markup for MDN annos for p-element-descendant
elements to be inserted after the p ancestor (</p> end tag). Without
this change, the annos otherwise get inserted before the p ancestor,
which is suboptimal because in many cases that results in the anno
overlapping with spec text (paragraph content). Inserting the annos
after the p ancestor minimizes the number of cases of such overlap.