MDN annos: Ensure headings annos go to right split
This change ensures that in multipage output, any MDN annotations associated
with heading elements/IDs show up in the same output file (split) as the
heading. Otherwise without this change, MDN annos for headings can end up
orphaned in the previous (split) output file.

This change also refines the source to eliminate code duplication, by rewriting
the code to have only one function for inserting the annotations, rather than
what previously had effectively been three functions with some redundant code.

And introduced here are new NextElementSibling() and PreviousElementSibling()
methods in the underlying DOM implementation, which when adding a new annotation
to a particular element are used for finding whether an annotation already
exists for that element — so the new annotation can be merged into the same
marginal box which holds the existing annotation.