MDN annos: Fix does-this-browser-support algorithm
This change rewrites the algorithm for determining from the BCD data
whether a particular browser supports a given feature.

Without this change, the existing code otherwise in some cases wrongly
causes multiple rows for some browsers to be added to the support tables
in the MDN annotations, with different version numbers and different/
conflicting indications of whether the browser supports the feature.