MDN annos: Fix "version_added: null" handling
This change fixes the handling for cases when the browser-compat-data/
mdn-spec-links JSON data for a spec feature includes a browser with a
version_added key that has the value "null". This change causes a row
with a '?' question to be shown for that browser, as expected.

Otherwise, without this change, no row would be output for that browser —
resulting in some annos having “missing” rows when compared to others.