MDN annos: Change markup to match Bikeshed annos
This change revises the element markup and attribute markup we generate
for the MDN annotations such that the markup matches what Bikeshed
generates. That allows the HTML standard to use the same MDN-annotations
CSS styles that we use for all other WHATWG standards.

Depends on https://github.com/whatwg/whatwg.org/pull/311