MDN annos: Fix placement of annos in tables/lists
This change ensures that for elements we’re annotating which are inside
tables or lists, each anno ends up placed as expected in horizontal
alignment with the element it’s annotating.

Otherwise, without this change, in cases where we have a long table (e.g.,
the list of events in the index) or long <dl> list (e.g., the list of
pseudo-classes), all the annos for everything in that table or list end up
being merged into a single annotation way up at the beginning of the table
or list — which prevents them from being discoverable to readers.