Build and deploy Wattsi on CI using Docker
This adds a Dockerfile which, when built, contains the Wattsi executable (with debian:stable-slim as the base). This effectively provides pre-built Wattsi binaries, uploaded to the Docker hub, for those with Docker available. It also gives us CI coverage for Wattsi, which we were lacking.

This closes #102 by removing the hard-coded version. With this in place the HTML build infrastructure should no longer need that.

Along the way, this adds shebangs to various .sh files (required in at least some OSes, including the Debian one we're using in Docker), removes the test commands from the compile.sh script (those can be run separately, after building), and consolidates build.sh into a Makefile.