Use multi-stage builds in the Dockerfile
This reduces the (uncompressed) size of the resulting Docker image from 99.3 MiB to 25.4 MiB, by starting a fresh image and only copying over the built result. A good chunk of this reduction comes from using the "distroless" base image for the final result (gcr.io/distroless/base).

By splitting the RUN commands, it also reduces iteration time: if the contents of /whatwg/wattsi/src change, running `make docker` will no longer require reinstalling all of the build dependencies inside the container, and instead the cached image with those dependencies will be reused, with only the steps from the COPY command onward re-done. (Splitting the RUN commands in this way was avoided previously because it would increase the size of the image, but with multi-stage builds, this is not a problem. See https://stackoverflow.com/a/61975952/3191.)