Fix handling of MDN annos for heading descendants
This change causes the handling of MDN annotations for descendants of
headings to match the existing handling we’re already using for headings
themselves (landed in dab062c). Specifically, if an element is a
descendant of a heading, any MDN annotation generated for that element
is inserted after the heading element in the DOM (rather than before).

Otherwise, without this change, if an element is a descendant of a
heading and we insert an MDN annotation for that element, the annotation
is inserted before the heading — which causes the same problem described
in the dab062c commit message (in multipage output, the annotation can
end up orphaned in the previous split; and in both multipage and
single-page output, the annotation isn’t aligned with the heading).