Add Document.GetDocType() & GetDocumentElement()
This change replaces the Document.DocType and Document.DocumentElement
properties with Document.GetDocType() and Document.GetDocumentElement()
methods in the src/html/dom.pas source, and updates the src/wattsi.pas
application code to use the Document.GetDocumentElement() method. (The
Document.DocType property was not actually used in the existing code.)

Otherwise, without that change, fpc fails with these fatal errors:

dom.pas(120,55) Error: Record or object type expected
dom.pas(121,58) Error: Record or object type expected

https://wiki.freepascal.org/User_Changes_Trunk#Property_field_access_lists_no_longer_allows_classes
explains that the remedy for those errors is to do one of the following:

* Switch the fields to records or objects
* Use a method with inline modifier that will result in similar
  performance

So this change uses the second of those remedies.