Adjust IGNORE_MESSAGES list based on fpc version
This change adjusts the list of numbered fpc (hint/warning) messages we
suppress (IGNORE_MESSAGES value in the src/lib/compile.sh script) based
on the fpc version.

Otherwise, without this change, if we give an older fpc version a
message number that it doesn’t know, that apparently causes fpc to fail.