Fix the <ref> processing again
The string/rope semantics continue to bite us. The original fix in bd1cd4460609a8284bb92fd4a4c407f4cc7b68fb was not correct, for unclear reasons. This version uses ExtractAll() to ensure we get a fresh CutRope that we can AppendDestructively() to Scratch, which is more in line with what other parts of the file do.