Allow Append()ing normal ShortStrings (length ≦ 255) to Ropes
This change reverts 0ad4342 and increases to 255 the length of strings
that can be passed directly to Rope≫Append() without getting truncated.
Otherwise, without this change, any string passed directly to
Rope≫Append() whose length exceeds a particular system-dependent limit
(which in practice seems to be 15) will get silently truncated.

Relates to #153.