Make the DOM implementation append strings directly
This change makes the DOM implementation append strings directly when
setting attribute values and when appending text nodes and comments.
That in turn allows passing non-constant strings to Element≫AppendChild()
and Document≫AppendChild() and Element≫SetAttribute().

Otherwise, without this change, the strings we use for attribute values and
text nodes and comments either must be constants, or else we have to create
CutRopes, append strings to them, and then use TText.CreateDestructively()
(to create text nodes and comments) and Element≫SetAttributeDestructively()
(to set attribute values).